# spec/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow — Projetos & Tarefas API
  version: 1.0.0
  description: |
    API for managing Projetos (Projects) and Tarefas (Tasks).

    ## Scope decisions
    - **No pagination.** List endpoints (`GET /projetos`, `GET /projetos/{id}/tarefas`)
      return the full matching collection as a single JSON array, in a
      defined order (`createdAt` ascending, `id` ascending as tie-breaker —
      see each operation's description). This is a deliberate scope
      decision for this challenge, not an oversight. Ordering is part of
      the contract, not an implementation detail left to an
      implementation's database defaults, and is covered by contract tests. A future
      revision may introduce `limit`/`cursor` query params without breaking
      existing clients, since the response is currently a bare array.
    - **No `DELETE /projetos`.** Deleting a project is out of scope for this
      challenge — a deliberate decision, not an oversight. Projects are
      deactivated via `PATCH /projetos/{id}` with `status: archived`, never
      removed. A future revision may add project deletion with its own
      cascade/authorization rules.
    - **Un-archiving is allowed.** `PATCH /projetos/{id}` may set `status`
      from `archived` back to `active`. This is deliberate: no invariant
      guards this direction (business rule 1 only guards *archiving*), and
      reactivating a project makes its tasks transitionable again — a
...
      others don't.
...

...

...
servers:
...
paths:
  /projetos:
    post:
      tags: [Projetos]
      summary: Create a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          headers:
            Location:
              required: true
              description: |
                URL of the newly created project, relative to the API's base
                path.
              schema:
                type: string
                format: uri-reference
                example: /projetos/3fa85f64-5717-4562-b3fc-2c963f66afa6
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ProjectBodyBadRequest"

    get:
      tags: [Projetos]
      summary: List projects
      operationId: listProjects
      description: |
        Returns every project matching the optional `status` filter,
        ordered by `createdAt` ascending, `id` ascending as tie-breaker.
        No pagination — see the API-level scope note in `info.description`.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter projects by status
          schema:
            $ref: "#/components/schemas/ProjectStatus"
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ProjectStatusFilterBadRequest"
  /projetos/{id}:
    parameters:
      - $ref: "#/components/parameters/ProjectIdParam"

    get:
      tags: [Projetos]
      summary: Get a project by id
      operationId: getProject
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/PathParamBadRequest"
        "404":
          $ref: "#/components/responses/ProjectNotFound"

    patch:
      tags: [Projetos]
      summary: Partially update a project
      operationId: updateProject
      description: |
        Patchable fields: `name`, `description`, `status`.

        Setting `status` to `archived` triggers business rule #1 below if the
        project still has a task `in_progress`. Resubmitting the project's
        current `status` is a no-op for that field only (`200`), not a
        transition; any other field submitted in the same request (e.g.
        `name`) still applies normally.

        A malformed path `id` is checked before the body — see the
        `400`-stage sub-order in `info.description`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/ProjectPatchBodyBadRequest"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
        "422":
          description: Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                archiveBlockedByInProgressTask:
                  $ref: "#/components/examples/ArchiveBlockedByInProgressTask"
  /projetos/{id}/tarefas:
    parameters:
      - $ref: "#/components/parameters/ProjectIdParam"

    post:
      tags: [Tarefas]
      summary: Create a task within a project
      operationId: createTask
      description: |
        Creating a task in a project whose `status` is `archived` triggers
        business rule #4 below.

        A malformed path `id` (the project's) is checked before the body —
        see the `400`-stage sub-order in `info.description`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          headers:
            Location:
              required: true
              description: |
                URL of the newly created task, relative to the API's base
                path.
              schema:
                type: string
                format: uri-reference
                example: /tarefas/9c858901-8a57-4791-81fe-4c455b099bc9
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/TaskCreateBodyBadRequest"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
        "422":
          description: Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                taskCreateInArchivedProject:
                  $ref: "#/components/examples/TaskCreateInArchivedProject"

    get:
      tags: [Tarefas]
      summary: List tasks of a project
      operationId: listTasks
      description: |
        Returns every task of the project matching the optional `status` and
        `priority` filters, ordered by `createdAt` ascending, `id` ascending
        as tie-breaker. When both filters are provided, a task must match
        both (AND semantics) to be included. No pagination — see the
        API-level scope note in `info.description`.

        A malformed path `id` is checked before the query parameters — see
        the `400`-stage sub-order in `info.description`.
      parameters:
        - name: status
          in: query
          required: false
          description: Filter tasks by status
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: priority
          in: query
          required: false
          description: Filter tasks by priority
          schema:
            $ref: "#/components/schemas/TaskPriority"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/TaskFiltersBadRequest"
        "404":
          $ref: "#/components/responses/ProjectNotFound"
...
components:
  parameters:
    ProjectIdParam:
      name: id
      in: path
      required: true
      description: |
        Project id. Canonical RFC 4122 textual form (8-4-4-4-12 hex
        digits) — non-canonical strings some platform parsers tolerate
        (e.g. `1-1-1-1-1`) are rejected with `400`; see the path-id scope
        decision in `info.description`. Hex digits are case-insensitive,
        matching RFC 4122's input rule — only the grouping/length shape
        is enforced.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
...
  responses:
    PathParamBadRequest:
      description: |
        The `id` path parameter is not a canonical RFC 4122 UUID
        (8-4-4-4-12 hex digits). This includes non-canonical forms that
        some platform parsers tolerate (e.g. `1-1-1-1-1`) — see the
        path-id scope decision in `info.description`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            invalidUuidPathParameter:
              $ref: "#/components/examples/InvalidUuidPathParameter"
            nonCanonicalUuidPathParameter:
              $ref: "#/components/examples/NonCanonicalUuidPathParameter"
    ProjectBodyBadRequest:
      description: |
        Malformed project creation body — missing or invalid `name`, or a
        field not accepted on create (e.g. `status`). Examples below are
        illustrative, not exhaustive — any schema violation (missing,
        invalid, oversized, unknown, or `readOnly` field) yields this same
        response shape, as does a body that cannot be parsed as JSON at
        all or whose value shapes cannot be bound to the declared types —
        see the unparseable/unbindable-body scope decision in
        `info.description`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
          examples:
            projectNameInvalid:
              $ref: "#/components/examples/ProjectNameInvalid"
            projectStatusSubmittedOnCreate:
              $ref: "#/components/examples/ProjectStatusSubmittedOnCreate"
            unparseableRequestBody:
              $ref: "#/components/examples/UnparseableRequestBody"
    ProjectPatchBodyBadRequest:
      description: |
        Malformed request for `PATCH /projetos/{id}`: invalid path
        parameter, invalid `name`, or an empty `{}` body. Examples below are
        illustrative, not exhaustive — any schema violation (missing,
        invalid, oversized, out-of-enum, unknown, or `readOnly` field)
        yields this same response shape, as does a body that cannot be
        parsed as JSON at all or whose value shapes cannot be bound to the
        declared types — see the unparseable/unbindable-body scope decision in
        `info.description`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
          examples:
            invalidPathParameterFieldError:
              $ref: "#/components/examples/InvalidPathParameterFieldError"
            projectNameInvalid:
              $ref: "#/components/examples/ProjectNameInvalid"
            emptyPatchBody:
              $ref: "#/components/examples/EmptyPatchBody"
            unparseableRequestBody:
              $ref: "#/components/examples/UnparseableRequestBody"
    TaskCreateBodyBadRequest:
      description: |
        Malformed request for `POST /projetos/{id}/tarefas`: invalid path
        parameter, invalid `title`, missing `priority`, or an out-of-enum
        `priority` value. Examples below are illustrative, not exhaustive —
        any schema violation (missing, invalid, oversized, out-of-enum,
        unknown, or `readOnly` field) yields this same response shape, as
        does a body that cannot be parsed as JSON at all or whose value
        shapes cannot be bound to the declared types — see the
        unparseable/unbindable-body scope decision in `info.description`.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
          examples:
            invalidPathParameterFieldError:
              $ref: "#/components/examples/InvalidPathParameterFieldError"
            taskTitleInvalid:
              $ref: "#/components/examples/TaskTitleInvalid"
            taskPriorityMissing:
              $ref: "#/components/examples/TaskPriorityMissing"
            taskPriorityInvalid:
              $ref: "#/components/examples/TaskPriorityInvalid"
            unparseableRequestBody:
              $ref: "#/components/examples/UnparseableRequestBody"
...
    ProjectStatusFilterBadRequest:
      description: |
        Malformed request for `GET /projetos`: the `status` query filter is
        outside the project status enum.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            projectStatusFilterInvalid:
              $ref: "#/components/examples/ProjectStatusFilterInvalid"
    TaskFiltersBadRequest:
      description: |
        Malformed request for `GET /projetos/{id}/tarefas`: invalid path
        parameter (including non-canonical UUID forms — see the path-id
        scope decision in `info.description`), or the `status`/`priority`
        query filter is outside its enum.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            invalidUuidPathParameter:
              $ref: "#/components/examples/InvalidUuidPathParameter"
            nonCanonicalUuidPathParameter:
              $ref: "#/components/examples/NonCanonicalUuidPathParameter"
            taskStatusFilterInvalid:
              $ref: "#/components/examples/TaskStatusFilterInvalid"
            taskPriorityFilterInvalid:
              $ref: "#/components/examples/TaskPriorityFilterInvalid"
    ProjectNotFound:
      description: Project not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            projectNotFound:
              $ref: "#/components/examples/ProjectNotFound"
...
  schemas:
    ProjectStatus:
      type: string
      enum: [active, archived]
    TaskStatus:
      type: string
      enum: [pending, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high]
    Project:
      type: object
      description: |
        Full project representation, as returned by the API. This response
        shape is distinct from the request shapes below — it includes
        server-assigned fields (`id`, `createdAt`) that clients never send.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          nullable: true
          maxLength: 2000
        status:
          $ref: "#/components/schemas/ProjectStatus"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: |
            UTC, `Z`-suffixed (RFC 3339, e.g. `2024-01-01T10:00:00Z`).
            Offsets are never emitted — see the datetime scope decision.
      required: [id, name, status, createdAt]
    CreateProjectRequest:
      type: object
      description: |
        Request body for `POST /projetos`. `status` is not accepted on
        create; every new project starts as `active`.
        `additionalProperties: false` means submitting `status`, `id`, or
        `createdAt` here is rejected with `400`.
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '.*\S.*'
        description:
          type: string
          nullable: true
          maxLength: 2000
      required: [name]
    UpdateProjectRequest:
      type: object
      description: |
        Request body for `PATCH /projetos/{id}`. Patchable fields only:
        `name`, `description`, `status`. `id` and `createdAt` are `readOnly`
        and any other field is unknown — both are rejected with `400`
        because of `additionalProperties: false`.
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '.*\S.*'
        description:
          type: string
          nullable: true
          maxLength: 2000
        status:
          $ref: "#/components/schemas/ProjectStatus"
    Task:
      type: object
      description: |
        Full task representation, as returned by the API. This response
        shape is distinct from the request shapes below — it includes
        server-assigned fields (`id`, `createdAt`, `completedAt`,
        `projectId`) that clients never send.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 2000
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: |
            UTC, `Z`-suffixed (RFC 3339, e.g. `2024-01-01T10:00:00Z`).
            Offsets are never emitted — see the datetime scope decision.
        completedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
          description: |
            UTC, `Z`-suffixed (RFC 3339, e.g. `2024-01-01T10:00:00Z`).
            Offsets are never emitted — see the datetime scope decision.
        projectId:
          type: string
          format: uuid
          readOnly: true
      required: [id, title, status, priority, createdAt, projectId]
    CreateTaskRequest:
      type: object
      description: |
        Request body for `POST /projetos/{id}/tarefas`. `status` is not
        accepted on create; every new task starts as `pending`.
        `additionalProperties: false` means submitting `status`,
        `completedAt`, `id`, `createdAt`, or `projectId` here is rejected
        with `400`. `priority` is required — there is no implicit default.
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '.*\S.*'
        description:
          type: string
          nullable: true
          maxLength: 2000
        priority:
          $ref: "#/components/schemas/TaskPriority"
      required: [title, priority]
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      properties:
        type:
          type: string
          format: uri
          description: |
            A stable URI identifying the specific problem type. It is an
            identifier, not necessarily a dereferenceable link.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code repeated for convenience.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: |
            URI reference identifying this specific occurrence of the
            problem — typically the relative request path.
      required: [type, title, status]
    ValidationProblemDetails:
      description: RFC 7807 Problem Details, extended with a per-field error list.
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Name of the offending field.
                  message:
                    type: string
                    description: Why the field is invalid.
                required: [field, message]
          required: [errors]
  examples:
    ProjectNotFound:
      summary: Projeto não encontrado
      value:
        type: https://taskflow.dev/errors/resource-not-found
        title: Recurso não encontrado
        status: 404
        detail: Nenhum projeto encontrado com id 3fa85f64-5717-4562-b3fc-2c963f66afa6
...
    ArchiveBlockedByInProgressTask:
      summary: Regra de negócio 1 — arquivamento bloqueado por tarefa em andamento
      value:
        type: https://taskflow.dev/errors/project-archive-blocked
        title: Projeto não pode ser arquivado
        status: 422
        detail: O projeto possui ao menos uma tarefa com status in_progress.
        instance: /projetos/3fa85f64-5717-4562-b3fc-2c963f66afa6
...
    TaskCreateInArchivedProject:
      summary: Regra de negócio 4 — criação de tarefa em projeto arquivado
      value:
        type: https://taskflow.dev/errors/task-create-project-archived
        title: Não é possível criar tarefa em projeto arquivado
        status: 422
        detail: O projeto 3fa85f64-5717-4562-b3fc-2c963f66afa6 está arquivado e não aceita novas tarefas.
        instance: /projetos/3fa85f64-5717-4562-b3fc-2c963f66afa6/tarefas
...
    ProjectNameInvalid:
      summary: Campo obrigatório ausente ou inválido (name)
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: Um ou mais campos do projeto são inválidos.
        errors:
          - field: name
            message: não pode estar em branco e deve ter no máximo 100 caracteres
    ProjectStatusSubmittedOnCreate:
      summary: Campo não aceito na criação (status)
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: O corpo da requisição contém um campo desconhecido ou não aceito na criação.
        errors:
          - field: status
            message: campo não aceito na criação; todo projeto novo inicia como active
    TaskTitleInvalid:
      summary: Campo obrigatório ausente ou inválido (title)
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: Um ou mais campos da tarefa são inválidos.
        errors:
          - field: title
            message: não pode estar em branco e deve ter no máximo 200 caracteres
    TaskPriorityMissing:
      summary: Campo obrigatório ausente (priority)
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: Um ou mais campos da tarefa são inválidos.
        errors:
          - field: priority
            message: "é obrigatório e deve ser um dos valores: low, medium, high"
    TaskPriorityInvalid:
      summary: Valor de enum inválido (priority)
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: Um ou mais campos da tarefa são inválidos.
        errors:
          - field: priority
            message: "deve ser um dos valores: low, medium, high"
...
    EmptyPatchBody:
      summary: Corpo do PATCH vazio
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: O corpo do PATCH não pode ser vazio; inclua ao menos um campo patchable.
        errors:
          - field: body
            message: deve conter ao menos uma propriedade patchable
    InvalidUuidPathParameter:
      summary: Identificador malformado no path
      value:
        type: https://taskflow.dev/errors/invalid-path-parameter
        title: Parâmetro de path inválido
        status: 400
        detail: O identificador informado não é um UUID válido.
    UnparseableRequestBody:
      summary: Corpo não parseável como JSON (sempre reportado como field body)
      value:
        type: https://taskflow.dev/errors/invalid-request-body
        title: Corpo da requisição inválido
        status: 400
        detail: O corpo da requisição não é um JSON válido.
        errors:
          - field: body
            message: JSON malformado
    NonCanonicalUuidPathParameter:
      summary: Identificador não-canônico (tolerado por parsers lenientes) também é rejeitado
      value:
        type: https://taskflow.dev/errors/invalid-path-parameter
        title: Parâmetro de path inválido
        status: 400
        detail: O identificador informado não é um UUID válido.
    InvalidPathParameterFieldError:
      summary: Identificador malformado no path (reportado como errors[])
      value:
        type: https://taskflow.dev/errors/invalid-path-parameter
        title: Parâmetro de path inválido
        status: 400
        detail: O identificador informado não é um UUID válido.
        errors:
          - field: id
            message: deve ser um UUID válido
    ProjectStatusFilterInvalid:
      summary: Valor de filtro status (projeto) inválido
      value:
        type: https://taskflow.dev/errors/invalid-query-parameter
        title: Parâmetro de query inválido
        status: 400
        detail: "O parâmetro status deve ser um dos valores: active, archived."
    TaskStatusFilterInvalid:
      summary: Valor de filtro status (tarefa) inválido
      value:
        type: https://taskflow.dev/errors/invalid-query-parameter
        title: Parâmetro de query inválido
        status: 400
        detail: "O parâmetro status deve ser um dos valores: pending, in_progress, done."
    TaskPriorityFilterInvalid:
      summary: Valor de filtro priority inválido
      value:
        type: https://taskflow.dev/errors/invalid-query-parameter
        title: Parâmetro de query inválido
        status: 400
        detail: "O parâmetro priority deve ser um dos valores: low, medium, high."